requirements-dev.txt
django-extensions==3.2.3
factory_boy==3.3.1
pre-commit == 3.8.0
pytest==8.3.4
pytest-django==4.9.0
pytest-factoryboy==2.7.0
